Implement partPercentage reset cases for all body parts
Backside body parts need a partPercentage reset case for when
walkingPercentage reaches 0.  This helps keep all of the separate body
parts in sync.

Frontside body parts need a partPercentage resetu case for when
partPercentage is greater than 1.0 because of imperfect increments during
the walkingPercentage stages.  Sometimes, a part's partPercentage never
increases beyond 1.0, but it is good to have this safehold in case we change
the increment sizes or increment durations for a body part.